move the generic parse item method down to rss parser
