+++ feedi/parsers/base.py
@@ -1,6 +1,4 @@
-import datetime
 import logging
-import traceback
 import urllib
 
 import favicon
@@ -49,52 +47,6 @@ class BaseParser:
         and return a (feed_metadata, [item_data]) tuple.
         """
         raise NotImplementedError
-
-    # FIXME move this logic to rss parser
-    def parse(self, entry, previous_fetch, skip_older_than):
-        """
-        Given an entry raw data (as produced by the `fetch` method) and parse
-        into a dictionary by using parse_* methods on each of the `cls.FIELDS`
-        names.
-        """
-        result = {}
-
-        try:
-            url = self.parse_entry_url(entry)
-            published = self.parse_remote_created(entry)
-            updated = self.parse_remote_updated(entry)
-        except Exception as error:
-            exc_desc_lines = traceback.format_exception_only(type(error), error)
-            exc_desc = ''.join(exc_desc_lines).rstrip()
-            logger.error("skipping errored entry %s %s",
-                         self.feed_name, exc_desc)
-            return
-
-        # don't try to process stuff that hasn't changed recently
-        if previous_fetch and updated < previous_fetch:
-            logger.debug('skipping up to date entry %s', entry.get('link'))
-            return
-
-        # or that is too old
-        if (skip_older_than and published and
-                datetime.datetime.utcnow() - published > datetime.timedelta(days=skip_older_than)):
-            logger.debug('skipping old entry %s %s', self.feed_name, url)
-            return
-
-        for field in self.FIELDS:
-            method = 'parse_' + field
-            try:
-                result[field] = getattr(self, method)(entry)
-            except Exception as error:
-                exc_desc_lines = traceback.format_exception_only(type(error), error)
-                exc_desc = ''.join(exc_desc_lines).rstrip()
-                logger.error("skipping errored entry %s %s %s",
-                             self.feed_name,
-                             url,
-                             exc_desc)
-                return
-
-        return result
 
     # TODO make this a proper cache of any sort of request, and cache all.
     def request(self, url):
